% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_heatmap.R
\name{CreateHeatmap}
\alias{CreateHeatmap}
\title{Create heatmaps on desired subsets of genes}
\usage{
CreateHeatmap(sig.genes.sam.obj, subsets.dir, file.format=c("geneid",
"symbol"))
}
\arguments{
\item{sig.genes.sam.obj}{Object returned from call to SigGenesSAM}

\item{subsets.dir}{Directory containing files of genes output from pathway 
analysis or simply genes of interest}

\item{file.format}{Indicator of how genes are identified in each file, be it
"geneid" or "symbol"}
}
\description{
Normalize input raw data using quantile and mloess methods. 
Plots of the normalized data along with a dendrogram clustering all samples 
will be stored in newly created pipeline directory.
}

