% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayHeatmap.R
\name{pathwayHeatmap}
\alias{pathwayHeatmap}
\title{Create heatmaps on desired subsets of genes}
\usage{
preprocessPlots(runSAM.obj, pathwaysDir, fileFormat=c("geneid",
"symbol"))
}
\arguments{
\item{runSAM.obj}{Object returned from call to runSAM}

\item{pathwaysDir}{Directory containing files of genes output from pathway 
analysis}

\item{fileFormat}{Indicator of how genes are identified in each file, be it
"geneid" or "symbol"}
}
\description{
Normalize input raw data using quantile and mloess methods. 
Plots of the normalized data along with a dendrogram clustering all samples 
will be stored in newly created pipeline directory.
}

