% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSAM.R
\name{runSAM}
\alias{runSAM}
\title{Identify significant genes through SAM}
\usage{
runSAM(backgroundSub.obj, response, delta)
}
\arguments{
\item{backgroundSub.obj}{Object returned from call to 
backgroundSub}

\item{response}{Vector of array group membership, 1=control, 2=experimental}

\item{delta}{Tuning parameter to obtain significant number of genes}
}
\value{
A list with components
\item{siggenesTable}{Combined data frame of genes having significant 
positive and negative correlation}
\item{dataCol}{Vector of column indices containing array data}
\item{ntext}{Number of leading text columns}
\item{response}{Vector of array group membership, 1=control, 2=experimental}
\item{pipelineName}{Name of pipeline generated from input file name sans 
extension}
\item{data}{Data frame of chosen normalization method data}
}
\description{
Implement SAM and compute significant genes given delta. Output
will consist of all significant genes ordered by increasing q-value and 
decreasing d-score.
}

