% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSimilarity.R
\name{sampleSimilarity}
\alias{sampleSimilarity}
\title{Generate plot of sample identify matrix and PCA}
\usage{
sampleSimilarity(runSAM.obj)
}
\arguments{
\item{runSAM.obj}{Object returned from call to runSAM}
}
\description{
Heatmap of sample similarity will be created along with a PCA
plot labelled with control and experimental groups.
}

