% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_genes_SAM.R
\name{SigGenesSAM}
\alias{SigGenesSAM}
\title{Identify significant genes through SAM}
\usage{
SigGenesSAM(background.subtraction.obj, class.compare.cols, 
class.compare.name, fdr.cutoff=0.1, response)
}
\arguments{
\item{background.subtraction.obj}{Object returned from call to 
BackgroundSubtraction}

\item{class.compare.cols}{Vector of column indices indicating which subset 
of arrays are to be compared for this comparison}

\item{class.compare.name}{String title given to the name of the comparison}

\item{fdr.cutoff}{Max FDR for SAM, will use delta value which results in max 
FDR below this cutoff}

\item{response}{For two class unpaired: vector of 1, 2 values that indicate 
group membership. For two class paired: vector of -1, 1, -2, 2, etc. 
values that indicate pairings.}
}
\value{
A list with components
\item{siggenes.table}{Combined data frame of genes having significant 
positive and negative correlation}
\item{data.col}{Vector of column indices containing array data}
\item{ntext}{Number of leading text columns}
\item{response}{Vector of array group membership, 1=control, 2=experimental}
\item{pipeline.name}{Name of pipeline generated from input file name sans 
extension}
\item{data}{Data frame of chosen normalization method data}
\item{class.compare.cols}{Value entered through class.compare.cols parameter}
\item{class.compare.name}{Value entered through class.compare.name parameter}
\item{symbol.index}{Column index that contains gene symbol}
}
\description{
Implement SAM and compute significant genes given delta. Output
will consist of all significant genes ordered by increasing q-value and 
decreasing d-score.
}

