% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backgroundCutoff.R
\name{backgroundCutoff}
\alias{backgroundCutoff}
\title{Visualize arrays to identify cutoff for background subtraction}
\usage{
backgroundCutoff(preprocessData.obj, method=c("mloess","quantile"), 
xlim.lo=0, xlim.hi)
}
\arguments{
\item{preprocessData.obj}{Object returned from call to 
preprocessData}

\item{method}{Type of normalization to use: "quantile" for quantile 
normalization; "mloess" for MLOESS normalization}

\item{xlim.lo}{Lower bound on X for histogram plot (binary logarithm)}

\item{xlim.hi}{Upper bound on X for histogram plot (binary logarithm)}
}
\description{
Plot the histogram of max gene illumination across all arrays
to identify cutoff for background subtraction. Upper and lower limits can be
changed to narrow focus. Plots will be saved in analysis pipeline 
directory.
}

