% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backgroundSub.R
\name{backgroundSub}
\alias{backgroundSub}
\title{Perform background subtraction on normalized data}
\usage{
backgroundSubtraction(initializePipeline.obj, 
method=c("mloess","quantile"), cutoff)
}
\arguments{
\item{initializePipeline.obj}{Object returned from call to 
initializePipeline}

\item{method}{Type of normalization to use: "quantile" or "q" for quantile 
normalization; "mloess" or "m" for mloess normalization}

\item{cutoff}{The cutoff value (binary logarithm) identified through
backgroundCutoff}
}
\value{
A list with components
\item{ntext}{Number of leading text columns}
\item{dataCol}{Vector of column indices containing array data}
\item{id}{Vector containing gene ID's}
\item{idInd}{Column index containing gene ID information}
\item{symbol}{Vector containing gene symbols}
\item{symbolIndex}{Column index containing gene symbol information}
\item{descStats}{Vector of column indices containing descriptive statistics}
\item{pipelineName}{Name of pipeline generated from input file name sans 
extension}
\item{data}{Data frame of descriptive stats and normalized data for chosen 
method}
}
\description{
Remove all genes that have an maximum intensity level below that
of the provided cutoff.
}

