% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_subtraction.R
\name{BackgroundSubtraction}
\alias{BackgroundSubtraction}
\title{Perform background subtraction on normalized data}
\usage{
BackgroundSubtraction(preprocess.data.obj, 
method=c("mloess", "quantile"), cutoff)
}
\arguments{
\item{preprocess.data.obj}{Object returned from call to 
preprocess.data}

\item{method}{Type of normalization to use: "quantile" for quantile 
normalization; "mloess" for MLOESS normalization}

\item{cutoff}{The cutoff value (binary logarithm) identified through
background.cutoff}
}
\value{
A list with components
\item{ntext}{Number of leading text columns}
\item{data.col}{Vector of column indices containing array data}
\item{id}{Vector containing gene ID's}
\item{id.ind}{Column index containing gene ID information}
\item{symbol}{Vector containing gene symbols}
\item{symbol.index}{Column index containing gene symbol information}
\item{desc.stats}{Vector of column indices containing descriptive statistics}
\item{pipeline.name}{Name of pipeline generated from input file name sans 
extension}
\item{normalized}{Data frame of descriptive stats and normalized data for 
the chosen method}
}
\description{
Remove all genes that have an maximum intensity level below that
of the provided cutoff.
}

