% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSAM.R
\name{runSAM}
\alias{runSAM}
\title{Identify significant genes through SAM}
\usage{
runSAM(backgroundSub.obj, classCompareCols, classCompareName, 
fdr.cutoff=0.1, response)
}
\arguments{
\item{backgroundSub.obj}{Object returned from call to 
backgroundSub}

\item{classCompareCols}{Vector of column indices indicating which subset 
of arrays are to be compared for this comparison}

\item{classCompareName}{String title given to the name of the comparison}

\item{fdr.cutoff}{Max FDR for SAM, will use delta value which results in max 
FDR below this cutoff}

\item{response}{Vector of 1, 2 values that indicate group membership}
}
\value{
A list with components
\item{siggenesTable}{Combined data frame of genes having significant 
positive and negative correlation}
\item{dataCol}{Vector of column indices containing array data}
\item{ntext}{Number of leading text columns}
\item{response}{Vector of array group membership, 1=control, 2=experimental}
\item{pipelineName}{Name of pipeline generated from input file name sans 
extension}
\item{data}{Data frame of chosen normalization method data}
\item{classCompareCols}{Value entered through classCompareCols parameter}
\item{classCompareName}{Value entered through classCompareName parameter}
\item{symbolIndex}{Column index that contains gene symbol}
}
\description{
Implement SAM and compute significant genes given delta. Output
will consist of all significant genes ordered by increasing q-value and 
decreasing d-score.
}

